Update Samiki Class (description)
- Update Samiki Constructor
- Create property value limitation [1]
- Add support fields [2]

[1]:
Since I don't want the Happiness or Hunger value to be greater than 100 or less than 0, I have defined special occasions.
If this property was been set with an incorrect value when instantiating the class, it'll correct the value

[2]:
I created these fields so that they can call its property (each field has its respective property) and define it avoiding the StackOverflow exception.